refactor(tui): split app/handle_key.rs into review/jump-popup/scroll dispatch
`rinkaku-tui/src/app/handle_key.rs` had grown to 1009 lines, over ADR
0028's 1000-line warn threshold. Splits it by responsibility, following
PR #229's own out-of-scope notes on this file: the review overlay's key
dispatch moves to app/review_key.rs, the jump popup's key dispatch moves
to app/jump.rs (alongside the jump-popup state it already owns), and the
viewport-height-aware scroll dispatch moves to app/scroll_key.rs. The
core handle_key dispatch (overlay priority chain + screen/focus match)
stays in app/handle_key.rs.

Pure code motion. No behavior change — every moved function body is
byte-identical to its original, verified by function-level diff. The
only textual changes are the two moved-and-now-cross-module-called
methods widening from private to pub(super), and module-doc updates
describing the new split.
